Move the `api-client` module under the `@bayou` namespace.
