+++ local-modules/@bayou/api-client/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "api-client",
+  "name": "@bayou/api-client",
   "version": "1.0.0",
 
   "dependencies": {
