+++ local-modules/doc-client/DocSession.js
@@ -2,7 +2,7 @@
 // Licensed AS IS and WITHOUT WARRANTY under the Apache License,
 // Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
 
-import { ApiClient } from 'api-client';
+import { ApiClient } from '@bayou/api-client';
 import { BaseKey } from '@bayou/api-common';
 import { TheModule as appCommon_TheModule } from '@bayou/app-common';
 import { Logger } from '@bayou/see-all';
